refactor: move the PropertyAccessor.php and relatives to the helper
