+++ EMS/helpers/src/PropertyAccess/PropertyAccessor.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace EMS\CommonBundle\Common\PropertyAccess;
+namespace EMS\Helpers\PropertyAccess;
 
 use EMS\Helpers\Standard\Base64;
 use EMS\Helpers\Standard\Json;
